refactor: api serializers to separate modules
