+++ src/api/views/user_image.py
@@ -10,7 +10,7 @@ from rest_framework.viewsets import GenericViewSet
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
-from api.serializers.user_page import UserImageSerializer
+from api.serializers.user_image import UserImageSerializer
 from user_preferences.models import UserPreferencesData
