+++ src/api/views/user_preferences_data.py
@@ -10,7 +10,7 @@ from rest_framework.viewsets import GenericViewSet
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
 
-from api.serializers.user_page import UserPreferencesDataSerializer
+from api.serializers.user_preferences_data import UserPreferencesDataSerializer
 from user_preferences.models import UserPreferencesData
